Highlight cells belonging to clues in the selected group
It's a bit of an edge case, but it's totally valid for a clue to belong
to many groups. An example is Guardian Cryptic 26330 - the cells for
clue 10a belong to both the 10a,18d group and the 15a,10a group. By
assocating the cell with exactly one group, the group encountered last
wins, so when selecting a cell in 10a, 15a is highlighted instead of 18d
which one would expect.

As a clue entry can only declare a single group, we should trust that it
describes the "most relevant" group to that clue --- ie. if the clue
appears at the start of a group then we can expect that group to be in
the entry's `group` field.

As such, a better way to find the cells to be highlighted is to test if
the cell's `clueIds` overlaps with the `selectedClue`'s `group`. Once we
do that, we no longer need to store a group for a cell, simplifying the
cell data model and initialisation process.